moving class methods to helper
